flaskapp/flaskapp/requirements.txt
click
dnspython
Flask
itsdangerous
Jinja2
MarkupSafe
passlib
pymongo
python-dotenv
Werkzeug
gunicorn
bcrypt
